refactor: clippy fixes
