+++ crates/papyrus_network/src/discovery/identify_impl.rs
@@ -1,7 +1,7 @@
 use libp2p::identify;
 
 use super::kad_impl::KadFromOtherBehaviourEvent;
-use crate::main_behaviour::mixed_behaviour;
+use crate::mixed_behaviour;
 
 pub const IDENTIFY_PROTOCOL_VERSION: &str = "/staknet/identify/0.1.0-rc.0";
