+++ crates/papyrus_rpc/src/rpc_metrics/rpc_metrics_test.rs
@@ -181,19 +181,19 @@ async fn server_metrics() {
         for line in metrics.split('\n').filter(|line| line.contains("V0_6")) {
             if line.contains("rpc_incoming_requests{method=\"blockNumber\"") {
                 println!("{}", line);
-                incoming_block_number = line.split(' ').last().unwrap().to_owned();
+                line.split(' ').last().unwrap().clone_into(&mut incoming_block_number);
             }
             if line.contains("rpc_failed_requests{method=\"blockNumber\"") {
                 println!("{}", line);
-                failing_block_number = line.split(' ').last().unwrap().to_owned();
+                line.split(' ').last().unwrap().clone_into(&mut failing_block_number);
             }
             if line.contains("rpc_incoming_requests{method=\"getStateUpdate\"") {
                 println!("{}", line);
-                incoming_get_state_update = line.split(' ').last().unwrap().to_owned();
+                line.split(' ').last().unwrap().clone_into(&mut incoming_get_state_update);
             }
             if line.contains("rpc_failed_requests{method=\"getStateUpdate\"") {
                 println!("{}", line);
-                failing_get_state_update = line.split(' ').last().unwrap().to_owned();
+                line.split(' ').last().unwrap().clone_into(&mut failing_get_state_update);
             }
         }
         (
